% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_covariate.R
\name{aggregate_covariate}
\alias{aggregate_covariate}
\title{Aggregate a covariate to polygon boundaries}
\usage{
aggregate_covariate(
  cov,
  template,
  agg_table,
  resample_method = "bilinear",
  ...
)
}
\arguments{
\item{cov}{(\link[terra:SpatRaster-class]{terra::SpatRaster}) Covariate to aggregate}

\item{template}{(\link[terra:SpatRaster-class]{terra::SpatRaster}) Template raster to use for aggregation}

\item{resample_method}{(character(1), default 'bilinear') Resampling method to use
for the covariate raster. See \code{\link[terra:resample]{terra::resample()}} for options.}

\item{...}{Additional arguments passed to \code{\link[mbg:aggregate_raster_to_polygons]{mbg::aggregate_raster_to_polygons()}}}
}
\description{
Aggregate a covariate to polygon boundaries
}
\details{
Quick population-weighted covariate aggregation using a template raster
}
\seealso{
\code{\link[mbg:aggregate_raster_to_polygons]{mbg::aggregate_raster_to_polygons()}}, \code{\link[terra:resample]{terra::resample()}}
}
