% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_three_panel_plot.R
\name{create_covariate_map}
\alias{create_covariate_map}
\title{Create a map of covariate values by health facility catchment}
\usage{
create_covariate_map(
  catchments_with_covs,
  district_bounds,
  outcome_colors,
  col_lims,
  log_scale,
  cov_label
)
}
\arguments{
\item{catchments_with_covs}{(\link[sf:sf]{sf::sf} object) Polygons merged with covariate data, must contain column \code{THIS_COV} for plotting}

\item{district_bounds}{(\link[sf:sf]{sf::sf} object) District boundaries to plot over catchment polygons}

\item{outcome_colors}{(\code{character(N)}) Vector of colors for filling the covariate map}

\item{col_lims}{(\code{numeric(2)}) Numeric vector of length two specifying color scale limits (min, max)}

\item{log_scale}{(\code{logical(1)}) Whether to use a log transformation for the color scale (\code{TRUE} or \code{FALSE})}

\item{cov_label}{(\code{character(1)}) Label for the covariate to use in the legend}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object showing the covariate map for catchments
}
\description{
Create a map of covariate values by health facility catchment
}
