% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_three_panel_plot.R
\name{create_three_panel_plot}
\alias{create_three_panel_plot}
\title{Create three-panel plots for a covariate}
\usage{
create_three_panel_plot(
  catchments_with_covs,
  cov_data,
  district_bounds,
  cov_field,
  viraemia_field,
  pop_field,
  plot_title,
  cov_label,
  hist_num_breaks = 30,
  outcome_colors = viridisLite::viridis(n = 100),
  col_lims = c(0, 1),
  log_scale = FALSE,
  vlines = NULL
)
}
\arguments{
\item{catchments_with_covs}{(\link[sf:sf]{sf::sf} object) Polygons merged with covariate data}

\item{cov_data}{(\link[data.table:data.table]{data.table::data.table}) Table containing at least fields with the
covariate data, viraemia, and population}

\item{district_bounds}{(\link[sf:sf]{sf::sf} object) District boundaries to plot over catchments}

\item{cov_field}{(\code{character(1)}) Name of the covariate field in data objects}

\item{viraemia_field}{(\code{character(1)}) Name of the viraemia field in \code{cov_data}}

\item{pop_field}{(\code{character(1)}) Name of the population field in \code{cov_data}}

\item{plot_title}{(\code{character(1)}) Title for the overall plot}

\item{cov_label}{(\code{character(1)}) Label for the covariate}

\item{hist_num_breaks}{(\code{integer(1)}, default 30) Number of breaks for the histogram}

\item{outcome_colors}{(\code{character(N)}, default \code{viridisLite::viridis(n = 100)}) Colors for
the outcome variable}

\item{col_lims}{(\code{numeric(2)}, default \code{c(0, 1)}) Color limits for the outcome variable}

\item{log_scale}{(\code{logical(1)}, default \code{FALSE}) Whether to plot the outcome variable
on a log scale}

\item{vlines}{(\code{data.frame}, default \code{NULL}) Data.frame with two fields: \code{label} and
\code{x}. Each row is a vertical line to add to the plot.}
}
\value{
Plots the three panel plot to the active graphics device, returns NULL
}
\description{
Create three-panel plots for a covariate
}
