% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_covariate.R
\name{discretize_covariate}
\alias{discretize_covariate}
\title{Discretize a covariate}
\usage{
discretize_covariate(vec)
}
\arguments{
\item{vec}{(\href{N}{numeric}) Vector of values}
}
\value{
Vector of discretized (0/1) values
}
\description{
Discretize a covariate
}
