% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_to_features.R
\name{distance_to_features}
\alias{distance_to_features}
\title{Calculate distance from each raster cell to vector features}
\usage{
distance_to_features(raster_template, vector, crs = "EPSG:4326")
}
\arguments{
\item{raster_template}{(\link[terra:SpatRaster-class]{terra::SpatRaster}) Raster template to calculate distance
across. Each non-NA cell in the template will be filled with a distance value}

\item{vector}{(\link[terra:SpatVector-class]{terra::SpatVector} or \link[sf:sf]{sf::sf}) Vector features to calculate distance
to.}

\item{crs}{(\link{character}, default 'EPSG:4326') CRS that will be used for measuring
distance. The raster template will be returned in the same CRS it was passed in.}
}
\value{
(\link[terra:SpatRaster-class]{terra::SpatRaster}) Raster filled with distance values.
}
\description{
Calculate distance from each raster cell to vector features
}
