% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district_plot.R
\name{district_plot}
\alias{district_plot}
\title{Function to plot a district map}
\usage{
district_plot(
  district_name,
  districts_sf,
  catchments_sf,
  catchment_colors,
  facility_data,
  out_fp
)
}
\arguments{
\item{district_name}{(\code{character(1)}) District name}

\item{districts_sf}{(\code{sf::sf}) Districts shapefile containing at least the field
"area_name"}

\item{catchments_sf}{(\code{sf::sf}) Catchments shapefile containing at least the field
"cluster_name"}

\item{catchment_colors}{(\code{character(N)}) Catchment color scheme for ggplot,
corresponding to the "cluster_name" field in \code{catchments_sf}}

\item{facility_data}{(\code{data.table::data.table}) Facility data including the following
fields:
\itemize{
\item facility_label (\code{character(1)}) Facility name
\item latitude (\code{numeric(1)}) Latitude
\item longitude (\code{numeric(1)}) Longitude
}}

\item{out_fp}{(\code{character(1)}) Output file path}
}
\value{
Returns NULL; saves district map to \code{out_fp}
}
\description{
Function to plot a district map
}
