% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhs_displacement.R
\name{draw_cluster_assignments}
\alias{draw_cluster_assignments}
\title{Draw cluster-to-zone assignments from posterior location probabilities}
\usage{
draw_cluster_assignments(probs, n_draws)
}
\arguments{
\item{probs}{(\code{data.table}) Output of \code{\link[=dhs_location_probabilities]{dhs_location_probabilities()}}, with fields
\code{cluster}, \code{zone}, \code{prob}. Clusters absent from \code{probs} are not returned.}

\item{n_draws}{(\code{integer(1)}) Number of imputations}
}
\value{
\code{data.table} with fields \code{draw}, \code{cluster}, \code{zone} (NA = outside all zones)
}
\description{
Draw cluster-to-zone assignments from posterior location probabilities
}
