% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district_plot.R
\name{letters_ref}
\alias{letters_ref}
\title{Set up letters to map up to 78 facilities}
\usage{
letters_ref()
}
\value{
(\code{character(1)}) Letters to map up to 78 facilities
}
\description{
Set up letters to map up to 78 facilities
}
