% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhs_paths.R
\name{normalize_district}
\alias{normalize_district}
\title{Normalise district names for matching DHS clusters to restriction polygons}
\usage{
normalize_district(x, fixes = NULL)
}
\arguments{
\item{x}{(\code{character(N)}) District names}

\item{fixes}{(\code{list}, default NULL) Named list mapping normalised names to
replacement normalised names}
}
\value{
Lower-case, letters-only names (\code{character(N)}) with fixes applied
}
\description{
Normalise district names for matching DHS clusters to restriction polygons
}
