% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district_plot.R
\name{short_label}
\alias{short_label}
\title{Function to shorten a facility name}
\usage{
short_label(facility_name)
}
\arguments{
\item{facility_name}{(\code{character(1)}) Facility name to shorten}
}
\value{
(\code{character(1)}) Shortened facility name
}
\description{
Function to shorten a facility name
}
