% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_covariate.R
\name{discretize_covariate}
\alias{discretize_covariate}
\title{Discretize a covariate}
\usage{
discretize_covariate(vec, num_jenks = 5)
}
\arguments{
\item{vec}{(\href{N}{numeric}) Vector of values}

\item{num_jenks}{(\href{1}{integer}) Number of Jenks natural breaks to use}
}
\value{
List with six items:
\itemize{
\item \code{quintiles}: Vector of quintile cutoffs
\item \code{q1}: Bottom quintile
\item \code{q5}: Top quintile
\item \code{jenks}: Jenks natural breaks classification
\item \code{j1}: Bottom Jenks class
\item \code{jtop}: Top Jenks class
}
}
\description{
Discretize a covariate
}
